Common GitHub (#20)
* Support reading secrets from AWS Secret Manager

Sometimes it is useful to instantiate a secret value from a centralized secrets storage like
Amazon Secrets Manager. To support that terraform-ci will read a secret from the secrets manager if a secret value in `.env/tf_env.json`
starts with `secretsmanager://`. That means terraform-ci needs to go to the secrets manager, extract the secret from the secret string or from the secret json.

The format of the URL is

```
secretsmanager://SECRET_ID[:JSON_KEY]
```

where `secretsmanager://` is an indicator that travis-ci needs to read the secret from the Secrets Manager,
`SECRET_ID` is a name of the secret. An optional `JSON_KEY` says the secret should be read from a JSON key.

Few valid examples:
```
secretsmanager:///path/to/secret:key
secretsmanager://secret:key
secretsmanager://secret
```
In two first the secret is a JSON. In the last one the secret value is a plain text string.

* Bump version: 0.6.2 → 0.7.0

* Update common variable

Some common variable like AWS_* and GITHUB_TOKEN are assigned from their TF_VAR_* conterparts.
If TF_VAR_github_token was instantiated from secrets manager GITHUB_TOKEN stored incorrect value (`secretsmanager://...` instead of the token itself).

* Bump version: 0.7.0 → 0.7.1